moved some functions to a new utils file
